requirements.txt
python-miio
requests
pycryptodome